PopupNavigatorの分割
